Update role to Leap 15.6 and podman
This commit is large, I apoligize for it but there was no clever way to
split this up. It's a long overdue maintenance update.

Update the molecule tasks to use podman instead of docker and update the
test runs to use Leap 15.5 and Leap 15.6.

Also do some cleanup of the tests and required changes to make them work
again.

Also remove Debian bullseye.

Signed-off-by: phoenix <felix.niederwanger@suse.com>